support for simple text support
