+++ pingpong/support.py
@@ -1,5 +1,8 @@
 from abc import ABC, abstractmethod
+from typing import Union, Literal
+from functools import cached_property
 
+from pydantic_settings import BaseSettings
 import aiohttp
 from discord import Webhook
 
@@ -32,3 +35,52 @@ class DiscordSupportDriver(BaseSupportDriver):
             msg += req.message
 
             await webhook.send(msg)
+
+
+class DiscordSettings(BaseSettings):
+    """Settings for getting help with Discord."""
+
+    type: Literal["discord"] = "discord"
+    webhook: str
+    invite: str
+    template: str = (
+        'We run a <a href="{invite}" class="underline" '
+        'rel="noopener noreferrer" target="_blank">'
+        "Discord server</a> where you can get help with PingPong."
+    )
+
+    def blurb(self) -> str:
+        return self.template.format(invite=self.invite)
+
+    @cached_property
+    def driver(self) -> DiscordSupportDriver:
+        return DiscordSupportDriver(self.webhook)
+
+
+class NoSupportSettings(BaseSettings):
+    type: Literal["none"] = "none"
+    placeholder: str = (
+        "We sadly cannot offer additional support for this app right now."
+    )
+
+    def blurb(self) -> str:
+        return self.placeholder
+
+    @cached_property
+    def driver(self) -> None:
+        return None
+
+
+class CustomTextSettings(BaseSettings):
+    type: Literal["text"] = "text"
+    text: str
+
+    def blurb(self) -> str:
+        return self.text
+
+    @cached_property
+    def driver(self) -> None:
+        return None
+
+
+SupportSettings = Union[DiscordSettings, NoSupportSettings, CustomTextSettings]
